adding employee from HCM

+++ src/main/java/com/metalsa/api/repository/LaborLevelRepository.java
@@ -3,5 +3,5 @@ package com.metalsa.api.repository;
 import com.metalsa.api.model.entity.LaborLevel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LaborLevelRepository extends JpaRepository<LaborLevel, Long> {
+public interface LaborLevelRepository extends JpaRepository<LaborLevel, Integer> {
 }
